refactor: I deleted/moved files to their respective directory and made the folder structure slimmer.
